fix to check yml is exsisted

+++ core/src/main/java/com/flowci/core/flow/FlowServiceImpl.java
@@ -83,11 +83,14 @@ public class FlowServiceImpl implements FlowService {
     @Override
     public Flow delete(String name) {
         Flow flow = get(name);
-        Yml yml = getYml(flow);
-
-        ymlDao.delete(yml);
         flowDao.delete(flow);
 
+        Yml yml = getYml(flow);
+        if (Objects.isNull(yml)) {
+            return flow;
+        }
+
+        ymlDao.delete(yml);
         return flow;
     }
